debug

+++ lzero/model/unizero_world_models/tokenizer.py
@@ -151,7 +151,7 @@ class Tokenizer(nn.Module):
         for i in labels.shape:
             total_dims *= i
         logits = logits.reshape(total_dims, -1)
-        labels = labels.reshape(total_dims)
+        labels = labels.reshape(total_dims).long()
         loss = F.cross_entropy(logits, labels)
         return loss
